requirements.txt
jsonschema
pyyaml
streamlit-antd-components
matplotlib
deepdiff